# packages/backend/src/api/openapi.yaml
openapi: 3.0.0
info:
  title: Scroblarr API
  version: 1.0.0
  description: |
    API documentation for Scroblarr - Media scrobbling service for Plex, Jellyfin, and Emby.

    This API allows you to manage users, sync media watching history across services, and configure integrations with TVTime and Trakt.
servers:
  - url: http://localhost:3000
    description: Development server
security:
  - bearerAuth: []
  - apiKeyAuth: []
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: Bearer token (Plex or Jellyfin access token)
    apiKeyAuth:
      type: apiKey
      in: header
      name: X-API-Key
      description: Global API key for API authentication
  schemas:
    Error:
      type: object
      required:
        - error
      properties:
        error:
          type: string
          description: Error message
        details:
          type: array
          items:
            type: object
          description: Validation error details (if applicable)
    Success:
      type: object
      required:
        - success
      properties:
        success:
          type: boolean
          example: true
        message:
          type: string
          nullable: true
...
    Settings:
      type: object
      properties:
        plexServerUrl:
          type: string
          nullable: true
          format: uri
        syncHistoryLimit:
          type: integer
          nullable: true
        jellyfinHost:
          type: string
          nullable: true
          format: uri
        jellyfinPort:
          type: integer
          nullable: true
        jellyfinUseSsl:
          type: boolean
          nullable: true
        jellyfinUrlBase:
          type: string
          nullable: true
        jellyfinApiKey:
          type: string
          nullable: true
        apiKey:
          type: string
          nullable: true
          description: Global API key for API authentication
    SettingsUpdate:
      type: object
      properties:
        plexServerUrl:
          type: string
          format: uri
        syncHistoryLimit:
          type: integer
          minimum: 10
          maximum: 10000
        jellyfinHost:
          type: string
          format: uri
        jellyfinPort:
          type: integer
          minimum: 1
          maximum: 65535
        jellyfinUseSsl:
          type: boolean
        jellyfinUrlBase:
          type: string
        jellyfinApiKey:
          type: string
        apiKey:
          type: string
          minLength: 1
          description: Global API key for API authentication
    SyncHistoryItem:
      type: object
      properties:
        id:
          type: string
        userId:
          type: string
        username:
          type: string
        mediaType:
          type: string
          enum: [episode, movie]
        mediaTitle:
          type: string
        source:
          type: string
          enum: [plex, jellyfin]
        tvdbEpisodeId:
          type: string
          nullable: true
        tvdbMovieId:
          type: string
          nullable: true
        imdbMovieId:
          type: string
          nullable: true
        imdbEpisodeId:
          type: string
          nullable: true
        tmdbMovieId:
          type: string
          nullable: true
        tmdbSeriesId:
          type: string
          nullable: true
        posterUrl:
          type: string
          nullable: true
        seasonNumber:
          type: integer
          nullable: true
        episodeNumber:
          type: integer
          nullable: true
        year:
          type: integer
          nullable: true
        success:
          type: boolean
        errorMessage:
          type: string
          nullable: true
        wasRewatched:
          type: boolean
        destinations:
          type: array
          items:
            type: string
          nullable: true
        syncedAt:
          type: string
          format: date-time
    SyncHistoryResponse:
      type: object
      properties:
        data:
          type: array
          items:
            $ref: '#/components/schemas/SyncHistoryItem'
        pagination:
          type: object
          properties:
            page:
              type: integer
            pageSize:
              type: integer
            total:
              type: integer
            totalPages:
              type: integer
    SyncStatistics:
      type: object
      properties:
        total:
          type: integer
        successful:
          type: integer
        failed:
          type: integer
        byMediaType:
          type: object
          properties:
            episode:
              type: integer
            movie:
              type: integer
        bySource:
          type: object
          properties:
            plex:
              type: integer
            jellyfin:
              type: integer
...
paths:
...
  /api/v1/settings:
    get:
      summary: Get all settings
      tags: [Settings]
      security:
        - bearerAuth: []
        - apiKeyAuth: []
      responses:
        '200':
          description: Settings
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Settings'
        '401':
          description: Unauthorized
        '403':
          description: Forbidden (not admin)
        '500':
          description: Internal server error

    patch:
      summary: Update settings
      tags: [Settings]
      security:
        - bearerAuth: []
        - apiKeyAuth: []
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/SettingsUpdate'
      responses:
        '200':
          description: Settings updated
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Settings'
        '400':
          description: Validation error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '401':
          description: Unauthorized
        '403':
          description: Forbidden (not admin)
        '500':
          description: Internal server error
  /api/v1/settings/plex:
    delete:
      summary: Remove Plex server configuration
      tags: [Settings]
      security:
        - bearerAuth: []
        - apiKeyAuth: []
      responses:
        '200':
          description: Configuration removed
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Settings'
        '401':
          description: Unauthorized
        '403':
          description: Forbidden (not admin)
        '500':
          description: Internal server error
  /api/v1/settings/jellyfin:
    delete:
      summary: Remove Jellyfin server configuration
      tags: [Settings]
      security:
        - bearerAuth: []
        - apiKeyAuth: []
      responses:
        '200':
          description: Configuration removed
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Settings'
        '401':
          description: Unauthorized
        '403':
          description: Forbidden (not admin)
        '500':
          description: Internal server error
  /api/v1/sync/history:
    get:
      summary: Get sync history
      tags: [Sync]
      security:
        - bearerAuth: []
      parameters:
        - name: page
          in: query
          schema:
            type: integer
            default: 1
            minimum: 1
        - name: pageSize
          in: query
          schema:
            type: integer
            default: 20
            minimum: 1
            maximum: 100
        - name: mediaType
          in: query
          schema:
            type: string
            enum: [episode, movie]
        - name: success
          in: query
          schema:
            type: boolean
        - name: sortBy
          in: query
          schema:
            type: string
            default: syncedAt
        - name: sortOrder
          in: query
          schema:
            type: string
            enum: [ASC, DESC]
            default: DESC
      responses:
        '200':
          description: Sync history
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SyncHistoryResponse'
        '400':
          description: Bad request
        '401':
          description: Unauthorized
        '500':
          description: Internal server error

    delete:
      summary: Delete sync history
      tags: [Sync]
      security:
        - bearerAuth: []
      requestBody:
        content:
          application/json:
            schema:
              type: object
              properties:
                ids:
                  type: array
                  items:
                    type: string
      responses:
        '200':
          description: History deleted
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                  deleted:
                    type: integer
                    nullable: true
        '401':
          description: Unauthorized
        '500':
          description: Internal server error
  /api/v1/sync/history/{id}:
    delete:
      summary: Delete sync history item
      tags: [Sync]
      security:
        - bearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      responses:
        '200':
          description: Item deleted
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Success'
        '401':
          description: Unauthorized
        '404':
          description: Item not found
        '500':
          description: Internal server error
  /api/v1/sync/statistics:
    get:
      summary: Get sync statistics
      tags: [Sync]
      security:
        - bearerAuth: []
      responses:
        '200':
          description: Statistics
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SyncStatistics'
        '401':
          description: Unauthorized
        '500':
          description: Internal server error
  /api/v1/sync/poster/{id}:
    get:
      summary: Get poster image for sync history item
      tags: [Sync]
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      responses:
        '200':
          description: Poster image
          content:
            image/jpeg: {}
            image/png: {}
        '302':
          description: Redirect to poster URL
        '403':
          description: Authentication required
        '404':
          description: Not found
        '500':
          description: Internal server error
...